third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/composite-expected.txt

This is a testharness.js-based test.
PASS Default value
PASS Change composite value
PASS Unspecified keyframe composite value when setting effect composite
FAIL Specified keyframe composite value when setting effect composite assert_equals: specified keyframe composite value should not be overridden by setting effect composite expected (string) "replace" but got (undefined) undefined
Harness: the test ran to completion.
